Move environ functions to separate file
